[FEAT] init popupView
